requirements.txt
# All for Python3

mysql-connector-python == 8.2.0
numpy == 1.24.3
boto3 == 1.28.77
pandas == 2.0.3
xgboost == 2.0.2
sklearn == 0.0.post5
seaborn == 0.12.2
pickleshare == 0.7.5
matplotlib == 3.8.0
python-dotenv == 1.0.0